Update cleaning scripts

+++ clean_vrpt_all.sh
@@ -29,7 +29,9 @@ WORK_DIR=.
 rm -rf $WORK_DIR/$PDBML_EXT $WORK_DIR/$VALID_INFO_ALT $WORK_DIR/$XML_VALID $WORK_DIR/$RDF_VALID $WORK_DIR/$MMCIF_VALID $WORK_DIR/$XML_VALID_ALT $WORK_DIR/$RDF_VALID_ALT $WORK_DIR/$MMCIF_VALID_ALT
 
 for dir in chk_sum_*; do
- cd $dir; find . -type f | xargs rm -f; cd ..; rmdir $dir
+ if [ $dir != "chk_sum_pdbml_cc" ] && [ $dir != "chk_sum_pdbml_noatom" ] ; then
+  cd $dir; find . -type f | xargs rm -f; cd ..; rmdir $dir
+ fi
 done
 
 WORK_DIR=test
